+++ tests/datasets/test_experiments.py
@@ -7,8 +7,8 @@ from phoenix.datasets.evaluators import (
     ConcisenessEvaluator,
     ContainsKeyword,
     HelpfulnessEvaluator,
+    create_evaluator,
 )
-from phoenix.datasets.evaluators.base import create_evaluator
 from phoenix.datasets.experiments import run_experiment
 from phoenix.datasets.types import (
     AnnotatorKind,
